+++ lib/client/jsx/utils/file.jsx
@@ -1,0 +1,3 @@
+export const fileKey = ({project_name, file_name}) => `${project_name}:${file_name}`;
+export const folderKey = ({project_name, folder_name}) => `${project_name}:${folder_name}`;
+
